criação do LocacaoController, DTOs e insercao de validacao nos inputs de classes de entrada

+++ src/main/java/com/example/imobiliaria/domain/model/Alugueis.java
@@ -4,6 +4,7 @@ import lombok.Data;
 import lombok.EqualsAndHashCode;
 
 import javax.persistence.*;
+import java.math.BigDecimal;
 import java.util.Date;
 
 @Data
@@ -20,7 +21,7 @@ public class Alugueis {
     @Temporal(TemporalType.DATE)
     private Date dt_vencimento;
 
-    private Double valor_pago;
+    private BigDecimal valor_pago;
     private String obs;
 
     @ManyToOne
